backend/sis_master/requirements.txt
Pillow
Flask
tensorflow>=2.0.0
torch # Có cần thêm vào không? Vì không thêm thì chạy báo lỗi
clip